Change format to blueprint format
