+++ db_extension.py
@@ -1,0 +1,3 @@
+from flaskext.mysql import MySQL
+
+mysql = MySQL()
